+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## !Deploy
+
+* Implemented `Convert-Currency` to perform currency conversions via the ExchangeRate-API Open Access API.
+* Implemented `Format-Currency` to decorate a value with a specified currency symbol.
+
 ## [0.0.0] - 2024-02-06
 
 * Initial commit